Created independent stencil file for w_fix_consrv_moment and updated transate test to test w_fix_consrv_moment
